Use Page.objects.best_match_for_path() to find Pages.
    * Adding feincms_page to the request turns out not to work
      due to the order in which middlewares and context processors
      are executed.
    * Using Page.objects.best_match_for_path() requires a lot of
      mocking in tests.  The import for FeinCMS's Page is inside
      the _get_path_for_page() method in the middleware class, so
      that we can avoid importing it in the tests.
